+++ just-scored/doc.go
@@ -1,0 +1,3 @@
+// Package justscored contains code necessary to add scored goals to
+// DynamoDB and notify a Slack channel about them
+package justscored