+++ just-scored/goals.go
@@ -1,13 +1,10 @@
-package main
+package justscored
 
 import (
 	"fmt"
 	"log"
-	"os"
 	"time"
 
-	"github.com/aws/aws-sdk-go/aws/session"
-	"github.com/guregu/dynamo"
 	"github.com/pawelad/just-scored/worldcup"
 )
 
@@ -67,31 +64,4 @@ func ParseMatchEvents(match *worldcup.Match) {
 			addGoal(&goal)
 		}
 	}
-}
-
-// addGoal adds passed goal to DynamoDB
-func addGoal(goal *Goal) (added bool, err error) {
-	db := dynamo.New(session.New())
-	table := db.Table(os.Getenv("DYNAMODB_TABLE"))
-
-	// Check if the goal is already in the DB
-	count, err := table.Get("EventID", goal.EventID).Count()
-	if err != nil {
-		log.Print(err)
-	}
-
-	if count != 0 {
-		log.Printf("Goal already added: '%+v'", goal)
-		return false, nil
-	}
-
-	log.Printf("Adding goal '%+v'", goal)
-	err = table.Put(goal).Run()
-
-	if err != nil {
-		log.Print(err)
-		return false, err
-	}
-
-	return true, nil
 }